requirements.txt
pandas>=1.5
numpy>=1.24
yfinance>=0.2.27
requests
beautifulsoup4
lxml
vaderSentiment
ta>=0.8.0
pyyaml
python-dateutil
tqdm
scikit-learn
torch          # later, for modelling
matplotlib
seaborn
shap>=0.44.1


# optional
# tradingeconomics (if used for CPI) or fredapi